requirements.txt
accelerate==1.7.0
bitsandbytes==0.45.5
datasets==3.6.0
deepspeed==0.16.9
diffusers==0.32.2
einops==0.6.1
flash_attn==2.6.2
httpx==0.23.3
huggingface_hub==0.31.4
numpy==1.26.4
omegaconf==2.3.0
openai==1.87.0 
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
packaging==25.0
Pillow==10.4.0
Requests==2.32.3
safetensors==0.5.3
setuptools==78.1.1
tabulate==0.9.0
timm==1.0.13
tokenizers==0.21.1
torch==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.52.1
xformers==0.0.29.post2
tenacity==8.3.0